Join and Stop accept session number from List.
`wemux list` outputs a list of all running sessions prefixed with a number.
`wemux join` and `wemux stop` can now be invoked with this number.
`wemux join 3` or `wemux j 3` would join the third server on the list.
`wemux stop` can now be passed either the number, or the session name.
If no argument is passed it will default to stopping the current session.